[backend] refactor: move shareable resources into a new module
