+++ apps/portal-api/src/modules/document/domain/document.domain.test.ts
@@ -12,16 +12,16 @@ import {
   LogicalOperator,
   OrderingMode,
 } from '../../../__generated__/resolvers-types';
-import { upsertConnectors } from '../../ingest-manifest/ingest-manifest.domain';
-import { ManifestInformation } from '../../ingest-manifest/ingest-manifest.model';
-import sampleExtractedManifest from '../../ingest-manifest/test/sample-extracted-manifest.json';
+import { upsertConnectors } from '../../shareable-resource/opencti/integration/ingest-manifest/ingest-manifest.domain';
+import { ManifestInformation } from '../../shareable-resource/opencti/integration/ingest-manifest/ingest-manifest.model';
+import sampleExtractedManifest from '../../shareable-resource/opencti/integration/ingest-manifest/test/sample-extracted-manifest.json';
 import {
   Connector,
   INTEGRATION_CONNECTOR_METADATA_KEYS,
   INTEGRATION_METADATA_KEYS,
   INTEGRATION_SERVICE_INSTANCE_ID,
   OPENCTI_INTEGRATION_DOCUMENT_TYPE,
-} from '../opencti/integrations/integrations.model';
+} from '../../shareable-resource/opencti/integration/integration.model';
 
 import { TestHelper } from '../../../../tests/test.helper';
 import { SERVICES, TEST_ORGANIZATIONS } from '../../../../tests/tests.const';
