+++ apps/portal-api/src/modules/shareable-resource/openaev/scenario/scenario.graphql
