+++ apps/portal-api/src/modules/shareable-resource/opencti/custom-dashboard/custom-dashboard.graphql
