+++ apps/portal-api/src/modules/shareable-resource/opencti/integration/ingest-manifest/ingest-manifest.domain.ts
@@ -1,18 +1,18 @@
 import {
   DocumentMetadataKeyCode,
   IntegrationType,
-} from '../../__generated__/resolvers-types';
-import { logApp } from '../../utils/app-logger.util';
-import { omit } from '../../utils/utils';
-import { DocumentApp } from '../document/document.app';
-import { DocumentDomain } from '../document/domain/document.domain';
+} from '../../../../../__generated__/resolvers-types';
+import { logApp } from '../../../../../utils/app-logger.util';
+import { omit } from '../../../../../utils/utils';
+import { DocumentApp } from '../../../../document/document.app';
+import { DocumentDomain } from '../../../../document/domain/document.domain';
+import { telemetryApp } from '../../../../telemetry/telemetry.app';
+import { buildCreateEvent } from '../../../../telemetry/telemetry.helper';
 import {
   Connector,
   INTEGRATION_CONNECTOR_METADATA_KEYS,
   OPENCTI_INTEGRATION_DOCUMENT_TYPE,
-} from '../document/opencti/integrations/integrations.model';
-import { telemetryApp } from '../telemetry/telemetry.app';
-import { buildCreateEvent } from '../telemetry/telemetry.helper';
+} from '../integration.model';
 import { base64ToUpload } from './ingest-manifest.helper';
 import { ManifestInformation } from './ingest-manifest.model';
