+++ apps/portal-api/src/modules/shareable-resource/opencti/integration/ingest-manifest/ingest-manifest.model.ts
@@ -1,4 +1,4 @@
-import { Connector } from '../document/opencti/integrations/integrations.model';
+import { Connector } from '../integration.model';
 
 export interface ManifestInformation extends Partial<Connector> {
   use_cases: string[];
