+++ apps/portal-api/src/modules/shareable-resource/opencti/integration/ingest-manifest/test/sample-extracted-manifest.json
