+++ apps/portal-api/src/modules/shareable-resource/opencti/integration/integration.graphql
