+++ apps/portal-api/src/modules/shareable-resource/opencti/integration/integration.resolver.ts
@@ -3,12 +3,12 @@ import {
   Resolvers,
 } from '../../../../__generated__/resolvers-types';
 import { logApp } from '../../../../utils/app-logger.util';
+import { DocumentChildrenDomain } from '../../../document/domain/document.children.domain';
+import { DocumentDomain } from '../../../document/domain/document.domain';
 import { getServiceInstance } from '../../../service/instance/service-instance.domain';
 import { subscriptionApp } from '../../../subscription/subscription.app';
 import { useCaseDomain } from '../../../use-case/use-case.domain';
-import { DocumentChildrenDomain } from '../../domain/document.children.domain';
-import { DocumentDomain } from '../../domain/document.domain';
-import { Integration } from './integrations.model';
+import { Integration } from './integration.model';
 
 const resolvers: Resolvers = {
   Integration: {
