+++ apps/portal-api/src/modules/telemetry/telemetry.app.test.ts
@@ -40,8 +40,8 @@ import type { ServiceInstanceId } from '../../model/kanel/public/ServiceInstance
 import { DocumentApp } from '../document/document.app';
 import { deleteDocuments } from '../document/document.helper';
 import * as DocumentUploadsHelper from '../document/document.uploads.helper';
-import { INTEGRATION_SERVICE_INSTANCE_ID } from '../document/opencti/integrations/integrations.model';
 import * as serviceInstanceDomain from '../service/instance/service-instance.domain';
+import { INTEGRATION_SERVICE_INSTANCE_ID } from '../shareable-resource/opencti/integration/integration.model';
 
 vi.mock('config', async (importOriginal) => {
   const mod = await importOriginal<{ default: typeof config }>();
