+++ apps/portal-api/src/server/initialize.helper.ts
@@ -23,13 +23,13 @@ import UserOrganization, {
   UserOrganizationId,
 } from '../model/kanel/public/UserOrganization';
 import UserOrganizationCapability from '../model/kanel/public/UserOrganizationCapability';
-import { IngestManifestApp } from '../modules/ingest-manifest/ingest-manifest.app';
 import {
   insertNewOrganization,
   loadOrganizationBy,
   updateOrganizationBy,
 } from '../modules/organization-management/organizations/organizations.domain';
 import { loadUserBy } from '../modules/organization-management/users/user-domain/users.domain';
+import { IngestManifestApp } from '../modules/shareable-resource/opencti/integration/ingest-manifest/ingest-manifest.app';
 import {
   ADMIN_UUID,
   PLATFORM_DOMAIN,
